dev-libs/libbsd: clang-17 versioned-symbols patch
When built with the LIBBSD_OVERLAY macro set, the following functions
are not built which results in undefined symbols in libbsd.map:

    vwarn, vwarnx, warn, warnx, verr, verrx, err, errx

leading to build failures similar to

    ld.lld: error: version script assignment of 'LIBBSD_0.10.0' to
    symbol 'vwarn' failed: symbol not defined

when linking with ld.lld since ld.lld does not allow undefined symbols
in the map by default.

Add guards, so these functions are excluded from libbsd.map when not
built.  Thereby, allowing ld.lld-16 and newer to link successfully with
default options. See
https://maskray.me/blog/2023-03-19-lld-16-elf-changes for ld.lld-16
change summary.

Bug: 915068

Signed-off-by: Nicholas Vinson <nvinson234@gmail.com>