media-gfx/waifu2x-ncnn-vulkan: fix build "with" glslang-1.3.275
Do not quite see why this is even looking for glslang libraries
at all with USE_SYSTEM_NCNN, there is nothing that use it that
I can see. System ncnn *does* use glslang on the other hand
(maybe it's trying to check if system ncnn has support? ...but
then RDEPEND already ensures this).

Meanwhile system ncnn itself (which is more recent) seems to
be handling current glslang properly.

Let's just yank out the entire block.

That aside, this still looks for glslangValidator, so the
dependency should instead be BDEPEND.

Closes: https://bugs.gentoo.org/922512
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>