net-p2p/qbittorrent: remove libiconv dependency
It was silently removed from 4.6.x as part of a bump. It was *added*
long ago in commit 6b33e72d8933aca72a03b23870178adc55ec3998 with the
informative and explanatory message:

> Add virtual/libiconv dep

I don't believe there's any valid justification for this. The
application makes heavy use of Qt, and that includes using Qt's string
translation features, which means it will use whatever Qt's APIs happen
to use internally. Nowhere in the source code is iconv mentioned at all.
If the virtual is needed then it is firmly in the realm of "citation
needed".

Remove it from 4.5 as well -- and the live ebuild.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>