net-p2p/qbittorrent: remove geoip dependency
It was silently removed from 4.6.x as part of a bump. Really, it could
be removed since a long time ago. Upstream dropped the use of a system
geoip database waaaaaaaaaaaaaaaay back when, in:

https://github.com/qbittorrent/qBittorrent/commit/79976fbfcee4b6da2bea7d3a09db4dddceb6faaf

It has been the case since 2015 and the 3.3.0 release, at least
(possibly longer but upstream has very odd git branch habits so it's
difficult to tell).

Ever since then the dependency has simply been dead weight.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>