dev-build/autoconf: skip AC_FUNC_GETLOADAVG test for 2.13
The macro itself is broken in autoconf-2.13 but this is caused by newer kernels
etc rather than anything else. Very, very few packages use autoconf-2.13, let
alone AC_FUNC_GETLOADAVG, so not going to worry about it much.

Do as Fedora and Debian do and skip the test for it.

Closes: https://bugs.gentoo.org/406723
Closes: https://bugs.gentoo.org/677926
Signed-off-by: Sam James <sam@gentoo.org>