dev-db/mysql: migrate checking for enabled lto, to tc-is-lto
This toolchain func was recently added, and vis a lot more reliable than
get-flagq. mysql attempts to manually handle some edge case, for example
if the active flags contain `-flto -fno-lto` then tc-is-lto gets it
correct but with is-flagq we have to manually check both. On the other
hand, mysql doesn't handle detecting -flto=8.

Using tc-is-lto is the preferred way to handle this from now on. We
would rather use this wherever possible.

Drops a bit of pointless logging. We don't need an `einfo` to tell us
that make.conf is being respected.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>