x11-drivers/nvidia-drivers: add 550.40.07 (unkeyworded beta)
No known issues, only unkeyworded due to its beta status, feel free
to try it. In theory should be fairly sane given most of its changes
were in 545 and this is likely due to become a production branch (and
a stable candidate here) next release.

Removes the need for a few things:
- kernel-6.7 patch (fixed)
- gcc14 patch (not *really* fixed, but conftests.sh uses -Wno-error)
- 6.6 kernel limitation for arm64, nvidia does a GPL test (untested)
- powerd amd64 limitation, files exist on arm64 too now (untested)

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>