dev-vcs/git-filter-repo: replace crude hack with upstream version hints
setuptools-scm has a number of ways to detect the version of a package.
Of course getting it from a git repo is the base approach. For pypi
tarballs, it detects the version from the PKG-INFO file shipped in
sdists.

We have neither. But setuptools-scm also supports manually overriding it
in cases where you *know* the answer. There is no need to create a dummy
file: just export the "PRETEND_VERSION" variable and setuptools-scm will
respect it as an override.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>