cargo.eclass: Output only the first create URI when PKGBUMPING
Modify cargo_set_crate_uris to output only the first crate URI when
PKGBUMPING is set for the current ebuild.  This makes pkgdiff-mg much
faster on Rust packages, as it does not have to fetch all the hundreds
of crates that aren't part of the diff anyway.  This is an improved
version of 92001837418f3a50e6571c0f533520b42f90d488, as the original
attempted not to output any URIs, and therefore could create empty
groups in SRC_URI that are invalid.

Signed-off-by: Michał Górny <mgorny@gentoo.org>