dev-libs/isa-l: keyword 2.30.0-r1 for -x86
Build failures. Test failures. Even if it did work, it would be rather
slow because the workaround for the former involves disabling
arch-optimised code paths. Upstream comments in the relevant issue
suggest they do not particularly care for 32-bit PCs.

Probably shouldn't have been keyworded for this arch in the first place.

Bug: https://bugs.gentoo.org/923425
Signed-off-by: Marek Szuba <marecki@gentoo.org>