media-libs/libaom: build without asserts by default
Both to follow upstream recommendations (see bug #921438 -- I did also have a better
source for this than the README but I can't find it now, or maybe it's changed
in the meantime), but also to avoid asserts firing during daily use which don't
seem to be bothering anybody upstream (bug #914614). Not ideal but it's been reported
already and went nowhere.

Closes: https://bugs.gentoo.org/914614
Closes: https://bugs.gentoo.org/921438
Signed-off-by: Sam James <sam@gentoo.org>