app-misc/ca-certificates: fix ewwy gross warning when no hooks exist
In the previous patch to update-ca-certificates, a continuation was
omitted for the case where no hooks exist, and POSIX sh expanded the
glob to "*". Bash has nullglob for this, but the POSIX solution is to
check if it is a file and skip otherwise.

The result: running update-ca-certificates logged an exit 127 of the
hook, then blithely continued and returned overall success since hooks
aren't defined as required to succeed.

Instead, we avoid running nonexistent hooks, which avoids logging any
return values and still continues while returning overall success, so,
no actual difference. But it does look nicer.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>