dev-util/yamllint: remove never-correct dependency on setuptools
This was erroneously introduced upstream in
https://github.com/adrienverge/yamllint/pull/321 with the logic that the
generated entrypoint script used pkg_resources to load the entrypoint.
This was always heavily conditional on how you installed it, and has
better distribution-wide approaches to handle it, and is anyways no
longer the case since a while now (because the code that generated
pkg_resources entrypoints has stopped doing so).

Upstream also accidentally got rid of this dependency while migrating to
pyproject.toml, indicating exactly how important it is.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>