profiles/arch/hppa: update and expand cups + pdf stable-mask
Previously, USE=pdf depended on mupdf, which is unstable for hppa and
blocked cups stabilization. This is still the case, except now the cups
packages have been reworked so there are additional packages exposing
the same USE=pdf, and which also need to be masked.

Add a couple extra masks to this group.

Bug: https://bugs.gentoo.org/923811
Bug: https://bugs.gentoo.org/627392
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>