profiles/arch/sparc: fix totally bogus mupdf-based cups mask
In bug 631970, mupdf dropped its sparc keyword for nonsense reasons. Due
to an ebuild bug, it was wrong everywhere and actually failed to build
on sparc. Despite diagnosis in the original report, it was left unfixed
for years after dropping ~sparc:

> Can someone with access to a sparc machine submit a patch? [...] if
> this bug report is correct, mupdf on sparc has now been broken for quite
> a while.  I'd tend to unkeyword it.

Fast forward to 2021, and commit 41543c0badfcd7ba9ee39386a3f4a8c8675135c0
which ninja-fixed this inside another commit. Now mupdf builds sanely.
Fast forward again to 2023, and bug 761550, where it was keyworded for
~sparc.

This odd mask is outdated and incorrect in a few different ways! The
status today is much more simple: the feature is expected to work on
~sparc, but no one has requested stabilizing mupdf for sparc. Hence,
cups *can* be built with USE=pdf, as long as you don't mind running
unstable keywords.

Move the mask from use, to use.stable, and correct the false
information.

Bug: https://bugs.gentoo.org/631970
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>