kernel-build.eclass: install vmlinu{x,z} symlink
Some other distributions install (a symlink to) the kernel
image here. Tools such as 'kernel-install list',
'kernel-install inspect', dracut and ukify look for the kernel
image here. So lets install this symlink to make manual
invocation of dracut and ukify a bit easier. As well as
make it possible to use other kernel-install features
such as list, inspect and add-all.

Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>