dev-qt/qttools: migrate live to llvm-r1, allow slot 18
Tested with 18.1.0_rc2 at same time, seems to be fine so may
as well do this now.

Skip .cmake.conf comment in 6.6.9999 given the minimum was
introduced in 6.7+. Unsure how often Qt intend to bump this,
odds are we may not need to pay attention to it if we clean
old versions up faster.

Not changing 6.6.1 given 6.6.2 is close and it can be updated
at same time to spare rebuilds.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>