dev-lang/ruby: remove ipv6 USE flag
Triggered by related discussion on gentoo-dev. The --enable-ipv6 option
only sets a default that is true on pretty much any system and tested in
any case, so it is a no-op for all Gentoo systems.

The lookup-order-hack mechanism is a workaround for issues in glibc that
were relevant 15 years ago and most likely already fixed there in the
mean time.

Signed-off-by: Hans de Graaff <graaff@gentoo.org>