app-editors/gvim: sync trivial details with vim ebuild
A lot of the logic between these packages is identical. And some of it
should be -- but because of things like variance between order of
pkg_postinst, whether or not to include "#" before a bug reference, or
capitalization in comments, the diff is larger than it has to be.

Larger diffs == harder to compare which changes are relevant. Make these
differences disappear by using consistent style between both.

It fixes style warts such as unsorted dependencies or IUSE as a bonus.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>