media-tv/kodi: Drop raspberry-pi USE requirement for libcec from all versions
This was previously only applied to 9999, but even the earlier libcec ebuild
from before the legacy support was dropped could be used against the new stack.
This makes more sense than adding (-) against this flag.

Strictly speaking, this should be done in revbumps, but we're going to have to
drop these older versions before we can drop the legacy Pi stack anyway.

Signed-off-by: James Le Cuirot <chewi@gentoo.org>