dev-db/influxdb: fix config handling
The config handling changed in v2 completly.
This means, the parameter "-config" is gone and does not exist anymore.

Instead, INFLUXD_CONFIG_PATH needs to be defined with the path,
where InfluxDB v2 needs to look for config files.

For compatibility reason, we will use '/etc/influxdb'.
This can be overriden by user.

Closes: https://bugs.gentoo.org/908437
Signed-off-by: Conrad Kostecki <conikost@gentoo.org>