x11-terms/kitty: workaround link failure with Go + gcc + -g(gdb)3
Not interested in pursuing a potential Go bug, so just do a
workaround. May recheck now and then to see if it's not longer
needed with toolchain updates.

If this is a widespread issue, it may need visiting in go-env's
go-env_set_compile_environment instead though.

Not an issue for the non-CGO_CFLAGS bits, so can keep -g3 elsewhere.

Closes: https://bugs.gentoo.org/924436
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>