dev-libs/mathjax: include USE=docs as part of the package (3.x)
Do not use a separated app-docs/mathjax-docs. It is impossible to get
correct: the symlink is generated using best_version, which applies at
build time but doesn't match the version at runtime, so having an -r1
breaks the package.

It also wasn't implemented correctly even if it could be gotten correct,
because the docs package is a build time dependency, not a runtime one.

To cap it off, this was introduced as a side effect of updating the
package to 3.x, in an attempt to solve the broken symlink in the 2.x
package. But it never referenced bug 867370 in the commit message.

While we are at it, fix a flaw in the code merged from mathjax-docs,
which didn't implement python_check_deps.

Fixes: 46560543c9d48506feb83de473a7b0bd0298ae83
Bug: https://bugs.gentoo.org/837722
Closes: https://bugs.gentoo.org/867370
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>