dev-libs/mathjax: include USE=docs as part of the package (2.x)
Do not use a separated app-docs/mathjax-docs. It is impossible to get
correct: the symlink is generated using ${SLOT}, which is totally bogus
as the docs package uses ${PF} and is unknowable. The symlink was very
broken.

In bug 526128, it was stated that there are 2 options for handling docs
as part of a separate tarball:

- remove doc support (and optionally create a separate package)
- make doc support depend on a separate package

It's baffling that the final option of just... having two SRC_URI
files and installing the docs... was totally overlooked.

Fixes: c6e861e430f795dd862fe05fed3f93ef2ca298a9
Bug: https://bugs.gentoo.org/526128
Bug: https://bugs.gentoo.org/867370
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>