dev-libs/mathjax: avoid inheriting vcs-clean
It does nothing. I am not sure what the point of using it here is --
this is a tarball source and doesn't have a .git/ directory. In 3.x it
does have a .github/ directory, but that isn't copied over even if it is
there.

It's a waste of time, so drop it.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>