app-text/calibre: add subslot dependency on mathjax
This doesn't often change. When it does, the calibre-server assets need
to be rebuilt with the new content hashes for the mathjax code.

Add the slot without a revbump. Any systems that don't pick up on this
update will be fine, because the chances are extremely high that calibre
will continue to be updated many times before the next mathjax update,
so no real point forcing a metadata-only rebuild.

Bug: https://bugs.gentoo.org/916037
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>