dev-qt/qtdeclarative: add missing qtbase[gui] dependency (qt6)
Exists in qt5's ebuild but seem to have went missing/overlooked
here, and noticed that this skips Quick entirely if gui is missing.

Albeit not really worth a revbump (slow rebuild), and disabling
gui is very unusual (the default IUSE=+widgets would also request
gui to be set on qtbase).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>