media-libs/noise-suppression-for-voice: Fix tests with gcc[-sanitize]
Upstream unconditionally adds -fsanitize=undefined to the test
executable target, but this is too unreliable to be worth the trouble
(e.g. doesn't work with gcc[-sanitize], and it probably causes other
issues).

Also fix running tests. The CMakeLists.txt configuration causes the
CTest file to end up in a subdirectory, which cmake_src_test doesn't
expect.

Also add USE="test", as the tests actually take a noticable amount of
time to compile compared to the rest of the files.

Closes: https://bugs.gentoo.org/924145
Signed-off-by: Kacper Słomiński <kacper.slominski72@gmail.com>
Closes: https://github.com/gentoo/gentoo/pull/35396
Signed-off-by: Sam James <sam@gentoo.org>