games-emulation/pcsx2: depend on libglvnd[X]
No need for a revbump given it does not *actually* link with
libGLX but looks for it during configure time.

It could be interesting to force it to be optional (or at least
not last time I tried, but given X libraries currently cannot be
optional either way (and upstream advise against using the wayland
support) it's likely not worth the extra trouble.

Will revisit if ever add a proper USE=X to pcsx2.

Bug: https://bugs.gentoo.org/924765
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>