dev-python/pythran: fix broken runtime metadata
Regression in commit 0c0d0150a89e320b064c09bfa5e6726bbd83a9da.

xsimd is a runtime dependency of pythran, by nature of what pythran is
and does. It defaults upstream to bundling it, but we remove that and
use the system package.

When performing that devendoring, it was mistakenly added as a build
time dependency (even though xsimd is not used at build time at all!)
but not a runtime dependency. This accidentally worked for people who
both installed pythran from source instead of using a binhost, and who
don't ever use `emerge -c --with-bdeps=n`. But it was very broken.

It is also logically used for the testsuite.

Revbump but don't drop keywords. It is a purely metadata-only fix and we
want that fix out ASAP.

Closes: https://bugs.gentoo.org/922782
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>