sys-devel/bison: Make gettext dependency optional
bison doesn't actually require gettext. The reason the dependency was
added unconditionally was because the configure script was automagically
adding the dependency if it was present at build time. This change adds
a proper --with-libtextstyle option to the configure script so that we
can control the dependency.

The patch was pushed upstream here:
https://lists.gnu.org/archive/html/bug-gnulib/2024-02/msg00190.html

Since the patch touched the m4 files we need to regenerate the configure
script. In order to avoid adding the autoconf dependency (via the
autotools eclass), I instead patch the configure script directly.

I wasn't sure if the timestamp stuff still applied, so I uncommented it
out of an abundance of caution.

Fixes: https://bugs.gentoo.org/925100
Signed-off-by: Raul E Rangel <rrangel@chromium.org>