app-arch/zstd: Fix cross build
zstd comes with the gen_html utility, used for generating
documentation. Meson configuration marks it as `native`, which
means building it natively even during cross builds. That was
done by developers on purpose[0] and it makes sense for manual
cross builds.

However, that doesn't work for cross builds with emerge. We need
to disable the `native` option on our own and the change is not
upstreamable.

[0] https://github.com/facebook/zstd/commit/132a1ad2915652aa86f32362c36a52e9b03fe695

Bug: https://bugs.gentoo.org/925287
Signed-off-by: Michal Rostecki <vadorovsky@protonmail.com>