app-crypt/gnupg: Fix cross-compiling by using gpgrt-config
It appears to be a reimplementation of pkg-config in Bash. It even uses the
SYSROOT properly without setting any environment variables, seemingly by basing
paths on where it was executed from. As such, using the build host's
gpgrt-config doesn't work. npth 1.7 is needed to cross-compile successfully but
older versions work for native builds.

Closes: https://bugs.gentoo.org/914711
Signed-off-by: James Le Cuirot <chewi@gentoo.org>