dev-util/buildbot-worker: add 3.11.1
Upstream point release officially drops dev-python/future as a
requirement by replacing it with dev-python/six instead, as a python2
compatibility library.

This is actually the only change for buildbot-worker, although there is
one commit each to buildbot and buildbot-www as well. This point release
truly is mild, like all the best bugfix patch releases.

Anyways, this is nice, because six is a tolerable dependency that we
have broad use of in the tree, whereas future is a hot mess, deprecated,
and broken entirely if you depend on it in python 3.12, so delayed
updating PYTHON_COMPAT in consumers.

Drop attempts at patching future out of buildbot-worker, it's not needed
anymore.

n.b. The new buildbot tag is not yet on PyPI, but buildbot gets
simultaneously released to Github Releases on a regular basis. Take this
opportunity to add the latter as a secondary download location.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>