dev-qt/qt-creator: disable designer by default in live
This been set by default in the old ebuilds and opted to leave it like
that but, with Qt6 and the non-split qttools where designer is off by
default, this been annoying for users.

Default designer on qttools when most users don't need it does not
seem like the right move -- and believe users that actually want this
will know to set it themselves (perhaps even globally).

Not that I'm either a qt-creator nor designer user (just maintaining
regardless), so opinions from actual users for defaults are welcome.

(live-only to avoid long rebuilds, will propagate in next version)

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>