sys-apps/debianutils: drop installkernel use flag
Random packages requiring some tool from Debian should not cause the kernel
installation process to change.

The dropping of the debianutils dependency in ca-certificates has already
caused some surprises due to installkernel being depcleaned. The origin of the
problem lies here in debianutils, users unknowingly use and rely on
installkernel but do not have it in their world file because it was implicitly
pulled in by some package that happens to use the run-parts command.

And also the other way around. If I am one of the users that wants to do
everything manually, I should not have my 'make install' unknowingly altered
by some package that I installed which pulled debianutils into the depgraph.

Drop this unused runtime dependency (which is against policy to begin with) and
its accompanying flag. A news item will be posted to notify users of this change
and instruct them to add installkernel to their world file.

Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>
Closes: https://github.com/gentoo/gentoo/pull/35533
Signed-off-by: Andrew Ammerlaan <andrewammerlaan@gentoo.org>