x11-drivers/nvidia-drivers: add 470.239.06
Also drop USE=wayland, last time I tried to use this with 470 not only
was it very limited due to the lack of GBM support, but it was starting
to break in horrible ways (e.g. flickering).

This been incompatible with newer egl-wayland, and 1.1.7 is filed with
problems. Users often find it annoying that portage warns that it
skipped the egl-wayland update too (USE=wayland is default on desktop
profiles even if they don't use it, so it was still annoying).

470 users should really just stick to Xorg (sorry).

Note nvidia-drm.modeset=1 was enabled with USE=wayland, disabling may
possibly have unexpected effects for some users even on Xorg. Re-enable
in nvidia.conf manually if you need this.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>