profiles/releases/23.0/package.use: re-enable USE=cli on mpv
Already default in the ebuild, but the profile's USE="-cli" causes
that to be overridden.

Hoping that this is a temporary workaround and there is a better way
to handle this in the profile, pretty sure the idea was to no longer
enable cli by default, not disable by default.

Bug: https://bugs.gentoo.org/925520
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>