sys-fs/jfsutils: mark as LTO-unsafe
Note: the filesystem is not a huge hit, and the utils package was last
released in 2011. It uses CVS as its version control, on sourceforge!
which means that it's been decommissioned in 2017 and there was no one
left to migrate or something. Discussion lists are inactive except the
one hooked into kernel.org, which is only used for discussing the kernel
driver not the utils.

It's not clear how to effectively report a utils issue, or what would
happen if we did.

Closes: https://bugs.gentoo.org/863905
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>