media-video/ffmpeg: fix crash when CPU_FLAGS_x86 and CFLAGS disagree
When CPU_FLAGS_X86 was set to the default for amd64 and CFLAGS was higher
with e.g. -march=native including AVX, we could get crashes because we disable
intrinsics which would otherwise adjust the alignment ffmpeg requests.

Backport an upstream patch for more sensible behaviour.

Note that the upstream patch doesn't apply to 4.4.x but I haven't tried
to backport it manually. Someone is welcome to, but 4.4.x should really be
on its way out by now.

Bug: https://trac.ffmpeg.org/ticket/10549
Bug: https://bugs.gentoo.org/915384
Signed-off-by: Sam James <sam@gentoo.org>