net-wireless/wpa_supplicant: call service stop on AP disconnect
Lot of discussion in the primary ticket for this, but apparently
multiple users report that previously connected wifi does not come up on
resume from suspend.  This is due to the original implementation
wpa_cli.sh not stopping the service, which was to avoid taking the
interface down, which would interfere with wpa_supplicant which needs
the interface to be up.  However, nowadays we set IN_BACKGROUND which
takes care of preventing the interface from being taken down, so it's
safe to stop the service.  If the service isn't stopped, then trying to
start the service on reconnection is a no-op (because the service is
already started).

Closes: https://bugs.gentoo.org/591752
Closes: https://bugs.gentoo.org/909574
Signed-off-by: Matoro Mahri <matoro_gentoo@matoro.tk>