net-wireless/gnuradio: remove erroneous dependency on mpir
mpir is an ancient fork of gmp from 2017. It claims to be focused on
speed. It doesn't build with modern compilers due to Modern C issues,
and it fails to build with LTO as well. Unlike gmp, this will never be
fixed.

gnuradio checks for "MPLIB", which is internally implemented by first
checking for gmp, and if gmp cannot be found, checking for mpir.

mpir was never required, and if gmp happened to be installed, it would
be an automagic dependency.

Furthermore, there was a USE-conditional dependency on gmp for wavelets,
which I am not sure makes sense but when triggered, resulted in a forced
gmp dependency that then guaranteed mpir wouldn't be used.

Bug: https://bugs.gentoo.org/812950
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>