dev-util/ruff: install as a standard cargo project
The python module has no functionality. It can be used solely as:

```
python3.11 -m ruff
```

in which case it will check for sys.prefix + '/bin/ruff' and try to run
it.

The downside of trying to install it as a python project is that it:
- adds fuzzy USE flags on python version
- recompiles everything once per python impl, despite attempts at
  copying sources around
- adds extra dependencies on e.g. maturin
- generally complicates the ebuild

We don't need any of that since this is just a cargo program.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>