distutils-r1.eclass: avoid confusing double equals notation in shell
Shell is a complex language to get consistently right already. The
double equals is almost-offensively redundant and pointless -- its
implementation is to be an exact alias for a single equals.

It causes muscle memory to be more likely to accidentally use this
bashism in #!/bin/sh scripts, it provides zero benefit, and the cherry
on top is that it takes up possibly-valuable real estate from your
screen width and causes (uncompressed) scripts to be (trivially) larger.

I wish Chet would repent of having added it and make bash issue a
bash-level warning on stderr if it encounters one in your scripts.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>