dev-python/shiboken6: disable terrible default of forcing limited API
The python Limited API allows you to compile a native extension once for
e.g. python 3.8, create a wheel package, and then install that on any
version of python 3.8+. This is inherently non-useful to distros, which
compile the extensions for each build regardless.

The Limited API is slower than the internal API, and benefits nothing at
all for our use case. But shiboken defaults to creating one. Manually
pass the option to disable it.

The Limited API builds of shiboken also have LTO issues that the regular
builds do not. Odd. See comments at https://bugreports.qt.io/browse/PYSIDE-2619

(Note that there are other LTO issues with shiboken which are fixed in
git dev, so we can't remove filter-lto yet.)

Bug: https://bugs.gentoo.org/925479
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>