x11-libs/xapp: fix broken, untested sed
This is why sed is bad. The sed didn't apply, even in the first edition
of the ebuild that it was introduced in, because it used incorrect sed
syntax. Because it was a sed, this fact escaped all notice even though
it never ever worked.

Just use a patch. Patches always either apply or fail, they aren't prone
to silently doing *nothing*.

Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>